requirements.txt
openai==2.21.0
instructor==1.14.5
python-dotenv==1.2.1
arize-phoenix == 12.*
openinference-instrumentation-openai == 0.1.*